+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseLinkApi.kt
@@ -1,0 +1,7 @@
+package com.affise.attribution.modules.link
+
+import com.affise.attribution.modules.AffiseModuleApi
+
+interface AffiseLinkApi : AffiseModuleApi {
+    fun linkResolve(url: String, callback: AffiseLinkCallback)
+}