+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseLinkCallback.kt
@@ -1,0 +1,7 @@
+package com.affise.attribution.modules.link
+
+
+fun interface AffiseLinkCallback {
+
+    fun handle(result: String)
+}