+++ attribution/src/main/java/com/affise/attribution/modules/subscription/AffiseSubscriptionApi.kt
@@ -1,18 +1,19 @@
 package com.affise.attribution.modules.subscription
 
 import android.app.Activity
+import com.affise.attribution.modules.AffiseModuleApi
 
-interface AffiseSubscriptionApi {
+interface AffiseSubscriptionApi : AffiseModuleApi {
     fun fetchProducts(
         productsIds: List<String>,
-        callback: AffiseResultCallback<AffiseProductsResult>
+        callback: AffiseResultCallback<AffiseProductsResult>,
     )
 
     fun purchase(
         activity: Activity,
         product: AffiseProduct,
         type: AffiseProductType?,
-        callback: AffiseResultCallback<AffisePurchasedInfo>
+        callback: AffiseResultCallback<AffisePurchasedInfo>,
     )
 
     fun purchase(
@@ -20,6 +21,6 @@ interface AffiseSubscriptionApi {
         productId: String,
         offerToken: String?,
         type: AffiseProductType?,
-        callback: AffiseResultCallback<AffisePurchasedInfo>
+        callback: AffiseResultCallback<AffisePurchasedInfo>,
     )
 }