+++ attribution/src/main/java/com/affise/attribution/network/HttpResponse.kt
@@ -5,4 +5,5 @@ data class HttpResponse(
     val code: Int,
     val message: String,
     val body: String? = null,
+    val headers: Map<String,List<String>> = emptyMap(),
 )