+++ module-link/src/main/java/com/affise/attribution/module/link/usecase/LinkResolveUseCase.kt
@@ -1,0 +1,7 @@
+package com.affise.attribution.module.link.usecase
+
+import com.affise.attribution.modules.link.AffiseLinkCallback
+
+internal interface LinkResolveUseCase {
+    fun linkResolve(url: String, callback: AffiseLinkCallback)
+}